users_service/requirements.txt
fastapi
uvicorn
pymongo
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator
python-multipart

